requirements.txt
cryptography==44.0.0
dj-database-url==2.3.0
Django==4.2.18
django-db-geventpool==4.0.1
django-cors-headers==4.6.0
django-picklefield==3.2
fyle-djangoq2==2.0.0
django-request-logging==0.7.5
djangorestframework==3.15.2
django-sendgrid-v5==1.2.0
future==1.0.0
fyle==0.37.2
fyle-accounting-mappings==1.36.3
fyle-integrations-platform-connector==1.39.3
fyle-rest-auth==1.7.2
gevent==24.11.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
requests==2.32.3
sageintacctsdk==1.26.0
sentry-sdk==2.19.2
six==1.17.0
Unidecode==1.1.2
sendgrid==6.11.0
python-dateutil==2.8.2
django-filter==24.3
pytz==2024.2

# Linting
flake8==7.1.1
flake8-docstrings==1.7.0
flake8-annotations==3.1.1